pyproject.toml
[project]
name = "tangle-cnn"
version = "0.98.0"
authors = [
  { name="Paul Yushkevich", email="pyushkevich@gmail.com" },
]
description = "Wildcat-based Weakly Supervised Segmentation for Digital Pathology in Neurodegenerative Diseases"
dependencies = [
    'numpy',
    'scipy',
    'torch>=1.6.0',
    'torchvision>=0.7.0',
    'matplotlib',
    'openslide-python',
    'SimpleITK',
    'parse',
    'Pillow'
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Development Status :: 4 - Beta",
    "Topic :: Scientific/Engineering :: Medical Science Apps."
]

[project.urls]
Homepage = "https://github.com/pyushkevich/tangle-cnn-prod"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
